Fix Unicode character compatibility in `lxc config edit`
Currently Unicode characters will become garbled when editing container
configs due to the unsupported user locale settings.  This patch fixes
the problem by resetting the locales to `C.UTF-8`.

Note that this is not a complete fix as it also disables the L10N as the
locales are hardcoded, the `locales-launch` remote part[1] is a solution
for this problem.

Fixes #29.

[1] https://forum.snapcraft.io/t/the-locales-launch-remote-part/8729

Signed-off-by: 林博仁(Buo-ren, Lin) <Buo.Ren.Lin@gmail.com>