Avoid implicit "." `source` dir in legacy Snapcraft
There is a default behavior of legacy 2.x Snapcraft which sets an
implicit "." `source` directory for parts without the `source` key.
This prevents parts that uses the `nil` plugin be considered "changed"
whenever a file under "." directory is modified.

This patch sets a fake source directory to all parts without the
`source` key to workaround the issue.

Refer the following snapcraft forum topic for more info:

    Snapcraft implicit source - snapcraft - snapcraft.io
    https://forum.snapcraft.io/t/snapcraft-implicit-source/7060

Signed-off-by: 林博仁(Buo-ren, Lin) <Buo.Ren.Lin@gmail.com>