build lxd binaries without using go plugin's build step
Perform the go build commands performed by the go plugin in
`override-build`.  The `override-build` is already very heavily
customized and there is little need to rely on the plugin anymore.

This patch simply adds a few extra build commands with the
correctly exported CGO_LDFLAGS to find the library dependencies
that are being built/staged.

Remove the custom go plugin, and rely on the `go` plugin again
but only using its `pull` functionality.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>