nogetgo: fix getting build base on snapcraft 4.0+
If available, use project._get_build_base() to get the project's
build base.

If _get_build_base() is not available (snapcraft < 4.0), then
resort to using project.info.base, which is safe to use for this
project as it specifies a base.  ProjectInfo will eventually be
removed, so this prioritizes the new interface.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>