snapcraft.yaml: use core22-step-dependencies
This build attribute insures that all parts are pulled during LP's
snapcraft pull call, and that they are not subsequently repulled again
during build.

This requires for pull steps of each part to not require the before
parts to be pulled.

This feature is currently present in snapcraft latest/edge only,
possibly will be available in 7.x/6.0.3.

Signed-off-by: Dimitri John Ledkov <dimitri.ledkov@canonical.com>