snapcraft: qemu: start shipping SeaBIOS 256k image
Let's start shipping a SeaBIOS 256k image in the snap.
It's needed for a CSM mode support in LXD.

Unfortunately, EDKII CSM module was removed:
https://bugzilla.tianocore.org/show_bug.cgi?id=4588

It means that for all new VMs in LXD we will start to use
SeaBIOS directly without EDKII/CSM module. For the compatibility
sake we are not removing EDKII-based CSM firmware right now,
but we will do that a little bit later.

Signed-off-by: Alexander Mikhalitsyn <aleksandr.mikhalitsyn@canonical.com>