snapcraft: remove vgabios for ati, cirrus and vmware GPUs
Those would require one to override the QEMU config and replace the virtio GPU
by one of those emulated devices. Situations where that would be useful are
probably very rare and many of the QEMU CVEs were due to those emulated
devices.

The goal is to limit the user's choice to a smaller list of hopefully safer
alternatives GPUs.

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>