snapcraft: make it clear that lz4 lib isn't primed
snapcraft deduplication is not applied to files from packages listed in
stage-packages.  To avoid duplicating lz4's .so, make it explicit that we don't
want to prime it and instead relying on the one provided in the base snap.

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>
(cherry picked from commit 2052ac1ccaafb7f3a1310cc6a5fd020347d9ba88)
Signed-off-by: Simon Deziel <simon.deziel@canonical.com>