snapcraft: replace usbutils by usb.ids
usbutils required usb.ids in Focal (core20) (when it was introduced in commit
ab0bc386594506b7b44) but that changed in Jammy (core22) where it no longer
pulled usb.ids by accident. Fortunately, we added a dependency on lshw which
does pull that in.

Add usb.ids to the list of dependencies makes it explicit that we want those.

Note: nothing from usbutils itself is present in 5.0/stable (5.0.2-d4d8da9)
snap.

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>