daemon.start: enhance how existing LXD PIDs are validated
By using some cheap random data to write and compare, we reduce the risk of a
nested LXD running the same code at the same time to also do this same
validation.

It is no longer needed to remove the .validate file first as we rely on its
content more than its presence.

Suggested-by: Alexander Mikhalitsyn <aleksandr.mikhalitsyn@canonical.com>
Signed-off-by: Simon Deziel <simon.deziel@canonical.com>