daemon.start: limit pgrep to match on root-owned LXD PIDs
This is a safeguard to ignore (unprivileged) nested LXD. Here is a comparison of the old way:

```
  server_pid: 1293
5923 /var/snap/lxd/common/lxd.debug --logfile /var/snap/lxd/common/lxd/logs/lxd.log --group lxd
8424 lxd --logfile /var/snap/lxd/common/lxd/logs/lxd.log --group lxd
```

In the above, the PID `8424` is the LXD inside the `c1` instance.

Now with the new way:

```
5923 /var/snap/lxd/common/lxd.debug --logfile /var/snap/lxd/common/lxd/logs/lxd.log --group lxd
```

Note: privileged nested LXD would also be running as EUID 0 but would be in a
different mount namespace which is what the `.validate` file comparison guards
against.

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>
(cherry picked from commit 5d7092d2fe8637a4c8ceab6fd691c0220c8b2f25)
Signed-off-by: Simon Deziel <simon.deziel@canonical.com>